requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alabaster==0.7.12
altair==4.2.0
aniso8601==9.0.1
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1659291887007/work
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
bqplot==0.12.36
branca==0.5.0
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click @ file:///Users/runner/miniforge3/conda-bld/click_1651215451259/work
click-plugins==1.1.1
cligj @ file:///home/conda/feedstock_root/build_artifacts/cligj_1633637764473/work
colour==0.1.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.17.1
earthengine-api==0.1.326
ee-extra==0.0.14
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
Fiona @ file:///Users/runner/miniforge3/conda-bld/fiona_1644294198451/work
Flask==2.1.3
Flask-Caching==2.0.1
Flask-Compress==1.13
Flask-Cors==3.0.10
flask-restx==0.5.1
folium==0.12.1.post1
fonttools==4.37.4
frozenlist==1.3.1
future==0.18.2
GDAL==3.4.1
gdown==4.5.1
geeadd==0.5.6
geemap==0.17.1
GeoAlchemy2==0.12.5
geocoder==1.38.1
geojson==2.5.0
geopandas==0.11.1
geovoronoi==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.1
google-api-python-client==1.12.11
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
h11==0.14.0
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.4
imagesize==1.4.1
importlib-metadata==4.12.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.16.0
ipyleaflet==0.17.1
ipysheet==0.5.0
ipython==8.5.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.7.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter-server-proxy==3.2.2
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.15.2
keplergl==0.3.2
kiwisolver==1.4.4
large-image==1.17.0
large-image-source-gdal==1.17.0
leafmap==0.10.5
localtileserver==0.6.0
logzero==1.7.0
lxml==4.9.1
mapclassify==2.4.3
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/a9eca864-069c-4240-73c8-4e51b036a10d/volume/mkl_random_1639994047876/work
mkl-service==2.4.0
multidict==6.0.2
munch==2.5.0
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nbserverproxy==0.8.8
nest-asyncio==1.5.6
networkx==2.8.6
notebook==6.4.12
notebook-shim==0.1.0
numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cer8_sm247/croots/recipe/numpy_and_numpy_base_1659432720132/work
OWSLib==0.27.2
packaging==21.3
palettable==3.3.0
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyCRS==1.0.2
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.4.0
pyrsistent==0.18.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.5.1
python-box==6.0.2
python-datauri==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-lorem==1.1.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
ratelim==0.1.6
requests==2.28.1
rich==12.5.1
rsa==4.9
sankee==0.2.0
scikit-learn==1.1.2
scipy==1.9.1
scooby==0.6.0
semver==2.13.0
Send2Trash==1.8.0
server-thread==0.2.0
Shapely @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_56b2jsagsp/croots/recipe/shapely_1664521598725/work
simpervisor==0.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.41
stack-data==0.5.1
streamlit==1.12.2
streamlit-bokeh-events==0.1.2
streamlit-folium==0.6.15
streamlit-keplergl==0.3.0
tenacity==8.1.0
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
traittypes==0.2.1
tropycal==0.6
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.1.2
whitebox==2.1.4
whiteboxgui==0.7.0
widgetsnbextension==3.6.1
xarray==2022.9.0
xyzservices==2022.9.0
yarl==1.8.1
zipp==3.8.1
